Refactor UniversalKey default action handling
- Replaced DefaultActionConfig with UniversalKeyDefaultActions
- Integrated GraphQL query for default actions data
- Enhanced action list to optionally hide header and delete functionality
- Updated UniversalKeyPage to use the new component
